added more ghost features, nullsound service and player death
The ghost is now able to scatter for 7 seconds and then chase for 20 seconds and loop this behavior.

The player can die and when she does, the ghosts will push a state to the stack of the state manager and play a interupting state that pauses the current state. then it sets itself as the currentstate.

the soundservice is now toggable between a working service and a null service that basically mutes the sound although if a sound is playing while muting it will first end that sound.

lives are now displayed on the bottom and the ghost touching the player decreases it and sends the player to the dying state.